feat: ability to set color of calendarItem
