+++ src/server/widgets/calendar/types.ts
@@ -1,10 +1,12 @@
 export type CalendarWidgetConfig = {
   url: string;
   daysInAdvance: number;
+  color: string;
 };
 
 export type CalendarWidgetData = {
   entries: CalendarEntry[][];
+  color: string;
 };
 
 export type CalendarEntry = {
